// src/app/models/sdac-keys.ts
export class SdacKeys {

    constructor() {
        this.active = '';
        this.activePubkey = '';
        this.basic = '';
        this.basicPubkey = '';
        this.memo = '';
        this.memoPubkey = '';
        this.owner = '';
        this.ownerPubkey = '';
    }

    active: string;
    activePubkey: string;
    basic: string;
    basicPubkey: string;
    memo: string;
    memoPubkey: string;
    owner: string;
    ownerPubkey: string;

}